[*] improve string table, don't user map for stridToStr in StringsItemDB

+++ Static/StringTable.cpp
@@ -93,7 +93,7 @@ bool matchSuffix(const std::string& searchStr, const UByteArrayAdapter & str) {
 	std::string::const_reverse_iterator searchStrIt = searchStr.rbegin();
 	std::string::const_reverse_iterator searchStrEnd = searchStr.rend();
 
-	uint16_t strPos = str.size()-1;
+	int32_t strPos = str.size()-1;
 	while (searchStrIt != searchStrEnd && strPos >= 0) {
 		if (static_cast<uint8_t>(*searchStrIt) != str.at(strPos))
 			return false;
